lms_backend/requirements.txt
# Django and related dependencies 
Django==5.1.3
djangorestframework==3.15.2
django-cors-headers==3.14.0  # If using CORS for API communication
